renamed systems.hpp and added test
