+++ include/dspbb/Filtering/IIR.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../LTISystems/DiscretizationTransforms.hpp"
-#include "../LTISystems/System.hpp"
+#include "../LTISystems/Systems.hpp"
 #include "IIR/BandTransforms.hpp"
 #include "IIR/Butterworth.hpp"
 #include "IIR/Descs.hpp"
