+++ include/dspbb/Filtering/IIR/Butterworth.hpp
@@ -2,7 +2,7 @@
 
 #include "../../Generators/Spaces.hpp"
 #include "../../Utility/Numbers.hpp"
-#include "../../LTISystems/System.hpp"
+#include "../../LTISystems/Systems.hpp"
 
 
 namespace dspbb {
