+++ include/dspbb/Filtering/IIR/Filter.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../LTISystems/System.hpp"
+#include "../../LTISystems/Systems.hpp"
 #include "Realizations.hpp"
